added friend feature

+++ MyHelper.Api/Services/MhTask/IMhTaskService.cs
@@ -8,9 +8,9 @@ namespace MyHelper.Api.Services.MHTask
 {
     public interface IMhTaskService
     {
-        Task<AOResult<List<MhTaskResponse>>> GetMhTasksAsync(long accountId, MhTaskFilterRequest mhTaskFilterRequest);
+        Task<AOResult<List<MhTaskResponse>>> GetMhTasksAsync(int accountId, MhTaskFilterRequest mhTaskFilterRequest);
 
-        Task<AOResult<MhTaskResponse>> GetMhTaskAsync(long accountId, long id);
+        Task<AOResult<MhTaskResponse>> GetMhTaskAsync(int accountId, long id);
 
         Task<AOResult> CreateMhTaskAsync(MhTaskRequest mhTaskRequest, MhTask parentMhTask = null);
